/src directory added
Added a header (it was supposed to be a footer like on the bottom of the page and have text-align: center; on the CSS... but I rage-quitted styling it since it's always sticking under an element) and I guess counts as a header since its under the header. Anyways, enjoy I guess?