Replaced PrettyOTA with ElegantOTA
Because apparently, the OTA library I used here before this update was taken down by DMCA due to copyright strike from the developers of ElegantOTA (Softt). Existing repositories with PrettyOTA linked or used are being flagged down and taken down by the DMCA in the request of the issuer of the copyright strike against LostInCompilation.